requirements.txt
torch>=2.1.0
transformers>=4.40.0
trl>=0.12.0
peft>=0.10.0
bitsandbytes>=0.43.0
datasets
accelerate
pyyaml
tqdm
vllm